requirements/base.txt
Django==5.2.9
django-filter==25.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg[binary,pool]
environs==14.5.0
python-swiftclient==4.9.0
django-storage-swift==1.4.0
celery==5.6.0
django-celery-beat==2.8.1
python-chrisstoreclient==1.0.0
python-pfconclient==4.0.0
django-auth-ldap==5.2.0
PyYAML==6.0.3
whitenoise[brotli]==6.11.0
PyJWT===2.10.1
channels==4.3.2
nats-py==2.12.0
uvicorn[standard]==0.38.0
git+https://github.com/tfranzel/drf-spectacular.git@refs/pull/1307/head
